bot/requirements.txt
aiogram==3.4.1
google-generativeai==0.3.2
openai-whisper==20231117
gtts==2.4.0
requests==2.31.0
python-dotenv==1.0.0
aiofiles==23.2.1
aiohttp==3.9.1
asyncio==3.4.3